Add docstrings to hybrid property, move MeasurableEntityType
