+++ app/db/utils.py
@@ -1,11 +1,19 @@
 import uuid
+from enum import StrEnum
 
 from pydantic import BaseModel
 from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import DeclarativeBase, InstrumentedAttribute, RelationshipProperty
 
-from app.db.model import Entity, Identifiable
+from app.db.model import Base, Entity, Identifiable, MeasurableEntity
 from app.logger import L
+
+MEASURABLE_ENTITIES: dict[str, type[MeasurableEntity]] = {
+    mapper.class_.__tablename__: mapper.class_
+    for mapper in Base.registry.mappers
+    if issubclass(mapper.class_, MeasurableEntity) and mapper.class_.__tablename__
+}
+MeasurableEntityType = StrEnum("MeasurableEntity", list(MEASURABLE_ENTITIES))
 
 
 def construct_model[T: DeclarativeBase](model_cls: type[T], data: dict) -> T:
