+++ app/filters/measurement_annotation.py
@@ -5,9 +5,9 @@ from fastapi_filter import FilterDepends, with_prefix
 
 from app.db.model import MeasurementAnnotation, MeasurementItem, MeasurementKind
 from app.db.types import MeasurementStatistic, MeasurementUnit, StructuralDomain
+from app.db.utils import MeasurableEntityType
 from app.filters.base import CustomFilter
 from app.filters.common import CreationFilterMixin
-from app.utils.entity import MeasurableEntityType
 
 
 class MeasurementItemFilter(CustomFilter):
