+++ app/schemas/measurement_annotation.py
@@ -4,8 +4,8 @@ from collections.abc import Sequence
 from pydantic import BaseModel, ConfigDict
 
 from app.db.types import MeasurementStatistic, MeasurementUnit, StructuralDomain
+from app.db.utils import MeasurableEntityType
 from app.schemas.base import CreationMixin, IdentifiableMixin
-from app.utils.entity import MeasurableEntityType
 
 
 class MeasurementItem(BaseModel):
